Further modifications for sampling.

+++ plugins/competence/parser/src/competenceparser.cpp
@@ -170,7 +170,7 @@ void CompetenceParser::traverseCommits(
     LOG(info) << "[competenceparser] Sampling " << allCommits << " commits in git repository.";
 
     // TODO: nice function to determine sample size
-    int sampleSize = std::sqrt((double)allCommits);
+    int sampleSize = (int)std::sqrt((double)allCommits);
     LOG(info) << "[competenceparser] Sample size is " << sampleSize << ".";
 
     RevWalkPtr walker2 = createRevWalk(repo.get());
